Changes on Search Input Field, loading element and Add the Navigation to TickectBooking Tab
