+++ package.json
@@ -15,7 +15,7 @@
     "axios": "^1.7.9",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^7.0.2",
+    "react-router-dom": "^7.1.1",
     "styled-components": "^6.1.13"
   },
   "devDependencies": {
