+++ src/Components/NavigationBar/PopularMovies.jsx
@@ -1,5 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
-import { useReducer } from "react";
+import { useNavigate } from "react-router-dom";
+import { useReducer, useState } from "react";
 import styled from "styled-components";
 
 const InputFields = styled.div`
@@ -10,6 +11,25 @@ const InputFields = styled.div`
   gap: 10px;
   margin-top: 10px;
 `;
+
+const Input = styled.input`
+  padding: 12px;
+  font-size: 16px;
+  width: 300px;
+  border: 2px solid #ddd;
+  border-radius: 5px;
+  transition: border-color 0.3s ease;
+  outline: none;
+
+  &:focus {
+    border-color: #007bff;
+  }
+
+  &::placeholder {
+    color: #888;
+  }
+`;
+
 
 const Button = styled.button`
   padding: 8px 16px;
@@ -47,45 +67,95 @@ const H3 = styled.h3`
 
 const Grid = styled.div`
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-  gap: 20px;
-  margin: 20px 0;
+  grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
+  gap: 30px;
+  margin-top: 20px;
+  padding: 20px;
+  box-sizing: border-box;
 `;
 
-const GridBlock = styled.div`
-  background-color: #f9f9f9;
+const GridBlock = styled.button`
+  background-color: #fff;
   border-radius: 8px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  overflow: hidden;
+  border: 1px solid #ddd;
+  padding: 15px;
   text-align: center;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.05);
+  transition:
+    transform 0.2s ease,
+    box-shadow 0.3s ease;
+
+  &:hover {
+    transform: translateY(-8px);
+    box-shadow: 0 6px 15px rgba(0, 0, 0, 0.1);
+  }
 `;
 
 const GridImg = styled.img`
   width: 100%;
   height: 300px;
   object-fit: cover;
+  border-radius: 4px;
+  margin-bottom: 15px;
+  transition: transform 0.3s ease;
+
+  &:hover {
+    transform: scale(1.05);
+  }
 `;
 
 const GridItem = styled.div`
-  padding: 10px;
-  font-size: 1rem;
-  color: #555;
+  font-size: 16px;
+  color: #333;
+  margin-top: 5px;
+
+  &:first-child {
+    font-weight: bold;
+    color: #007bff;
+  }
 `;
 
+
+
+const pagereducer = (pageNumber, action) => {
+  switch (action.type) {
+    case "nextPage":
+      return { count: pageNumber.count + 1 };
+    case "previousPage":
+      return { count: pageNumber.count > 1 ? pageNumber.count - 1 : 1 };
+    default:
+      return pageNumber;
+  }
+};
+
 const PopularMovies = () => {
-  const pagereducer = (pageNumber, action) => {
-    switch (action.type) {
-      case "nextPage":
-        return { count: pageNumber.count + 1 };
-      case "previousPage":
-        return { count: pageNumber.count > 1 ? pageNumber.count - 1 : 1 };
-      default:
-        return pageNumber;
+  const [pageNumber, dispatchPageNumber] = useReducer(pagereducer, {
+    count: 1
+  });
+
+  const navigate = useNavigate();
+
+  const [moviename, setMoviename] = useState("");
+
+  const options = {
+    method: "GET",
+    headers: {
+      accept: "application/json",
+      Authorization:
+        "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJkNTUyZWMyMGZlOWUxYTkzMzIzOTQwNzFmMzg2YTNmOCIsIm5iZiI6MTczNDc1MjI1Ny4xMzQsInN1YiI6IjY3NjYzODAxNmNlYmE4MjliOTc0YjQyMiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.s8XtBP1-lD9E6BgnaruBBzKWU92bQI_weSQNhDvX7a8"
     }
   };
 
-  const [pageNumber, dispatchPageNumber] = useReducer(pagereducer, {
-    count: 1
+  const fetchMovies = async () => {
+    const url = `https://api.themoviedb.org/3/search/movie?query=${moviename}&include_adult=false&language=en-US&page=1`;
+    const fetchedmoviesdata = await fetch(url, options);
+    return await fetchedmoviesdata.json();
+  };
+
+  const movielistsearch = useQuery({
+    queryKey: ["movie", moviename],
+    queryFn: fetchMovies,
+    enabled: moviename !== ""
   });
 
   const fetchPopularMovies = async () => {
@@ -95,7 +165,7 @@ const PopularMovies = () => {
       headers: {
         accept: "application/json",
         Authorization:
-          "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJkNTUyZWMyMGZlOWUxYTkzMzIzOTQwNzFmMzg2YTNmOCIsIm5iZiI6MTczNDc1MjI1Ny4xMzQsInN1YiI6IjY3NjYzODAxNmNlYmE4MjliOTc0YjQyMiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.s8XtBP1-lD9E6BgnaruBBzKWU92bQI_weSQNhDvX7a8" // Replace 'YOUR_API_KEY' with your actual API key
+          "Bearer eyJhbGciOiJIUzI1NiJ9.eyJhdWQiOiJkNTUyZWMyMGZlOWUxYTkzMzIzOTQwNzFmMzg2YTNmOCIsIm5iZiI6MTczNDc1MjI1Ny4xMzQsInN1YiI6IjY3NjYzODAxNmNlYmE4MjliOTc0YjQyMiIsInNjb3BlcyI6WyJhcGlfcmVhZCJdLCJ2ZXJzaW9uIjoxfQ.s8XtBP1-lD9E6BgnaruBBzKWU92bQI_weSQNhDvX7a8"
       }
     };
 
@@ -111,67 +181,110 @@ const PopularMovies = () => {
   const movielist = useQuery({
     queryKey: ["movie", pageNumber.count],
     queryFn: fetchPopularMovies,
-    enabled: true
+    enabled: !moviename
   });
 
   return (
     <>
       <InputFields>
-        <Button
-          onClick={() => {
-            dispatchPageNumber({ type: "previousPage" });
+        <label htmlFor="searchText"> Movie Title : </label>
+        <Input
+          type="text"
+          defaultValue={moviename}
+          onKeyDown={(e) => {
+            if (e.key === "Enter") setMoviename(e.target.value);
           }}
-          disabled={pageNumber.count === 1}
-        >
-          Previous Page
-        </Button>
-        <Button
-          onClick={() => {
-            dispatchPageNumber({ type: "nextPage" });
-          }}
-        >
-          Next Page
-        </Button>
+          placeholder="movie name"
+          id="searchText"
+        />
+        {moviename !== "" && movielistsearch.isLoading && <div class="loader"></div>}
       </InputFields>
 
-      <OutputField>
-        {/* Loading state */}
-        {movielist.isLoading && <p>Loading Popular Movies...</p>}
-
-        {/* Error state */}
-        {movielist.error && <p>Error: {movielist.error.message}</p>}
-
-        {/* Popular Movies Results */}
-        {movielist.data &&
-        movielist.data.results &&
-        movielist.data.results.length > 0 ? (
-          <>
-            <H3>Popular Movies</H3>
-            <Grid>
-              {movielist.data.results.map((movie) => (
-                <GridBlock key={movie.id}>
-                  <GridImg
-                    src={
-                      movie.poster_path
-                        ? `https://image.tmdb.org/t/p/w200${movie.poster_path}`
-                        : "https://via.placeholder.com/200x300?text=No+Image"
-                    }
-                    alt={movie.title}
-                  />
-                  <GridItem>{movie.title}</GridItem>
-                  <GridItem>
-                    {movie.release_date
-                      ? movie.release_date.split("-")[0]
-                      : "N/A"}
-                  </GridItem>
-                </GridBlock>
-              ))}
-            </Grid>
-          </>
+      {moviename !== "" ? (
+        movielistsearch.error ? (
+          <p>Error: {movielistsearch.error.message}</p>
+        ) : movielistsearch.data &&
+          movielistsearch.data.results &&
+          movielistsearch.data.results.length > 0 ? (
+          <Grid>
+            {movielistsearch.data.results.map((movie) => (
+              <GridBlock key={movie.id}>
+                <GridImg
+                  src={
+                    movie.poster_path
+                      ? `https://image.tmdb.org/t/p/w200${movie.poster_path}`
+                      : `https://via.placeholder.com/200x300?text=No+Image`
+                  }
+                  alt={movie.title}
+                />
+                <GridItem key={movie.imdbID}>{movie.title}</GridItem>
+                <GridItem>{movie.release_date}</GridItem>
+              </GridBlock>
+            ))}
+          </Grid>
         ) : (
-          !movielist.isLoading && <p>No popular movies found.</p>
-        )}
-      </OutputField>
+          !movielistsearch.isLoading && <p>No movies found.</p>
+        )
+      ) : (
+        <>
+          <InputFields>
+            <Button
+              onClick={() => {
+                dispatchPageNumber({ type: "previousPage" });
+              }}
+              disabled={pageNumber.count === 1}
+            >
+              Previous Page
+            </Button>
+            <Button
+              onClick={() => {
+                dispatchPageNumber({ type: "nextPage" });
+              }}
+            >
+              Next Page
+            </Button>
+          </InputFields>
+
+          <OutputField>
+            {movielist.data &&
+            movielist.data.results &&
+            movielist.data.results.length > 0 ? (
+              <>
+                <H3>Popular Movies</H3>
+                <Grid>
+                  {movielist.data.results.map((movie) => (
+                    <GridBlock
+                      key={movie.id}
+                      onClick={() => {
+                        navigate("/TicketBooking", {
+                          state: { movieid: `${movie.id}` }
+                        });
+                      }}
+                    >
+                      <GridImg
+                        src={
+                          movie.poster_path
+                            ? `https://image.tmdb.org/t/p/w200${movie.poster_path}`
+                            : "https://via.placeholder.com/200x300?text=No+Image"
+                        }
+                        alt={movie.title}
+                      />
+                      <GridItem>{movie.title}</GridItem>
+                      <GridItem>
+                        {movie.release_date
+                          ? movie.release_date.split("-")[0]
+                          : "N/A"}
+                      </GridItem>
+                    </GridBlock>
+                  ))}
+                </Grid>
+              </>
+            ) : (
+              !movielist.isLoading && <p>No popular movies found.</p>
+            )}
+          </OutputField>
+        </>
+      )}
     </>
   );
 };
